+++ app/src/main/java/org/linphone/activities/DialerActivity.java
@@ -24,6 +24,7 @@ import android.content.Intent;
 import android.content.res.Configuration;
 import android.os.Bundle;
 import android.view.LayoutInflater;
+import android.view.TextureView;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
@@ -171,15 +172,46 @@ public class DialerActivity extends MainActivity implements AddressText.AddressC
         }
 
         updateLayout();
+        enableVideoPreviewIfTablet(true);
     }
 
     @Override
     protected void onPause() {
-        super.onPause();
-
+        enableVideoPreviewIfTablet(false);
         Core core = LinphoneManager.getCore();
         if (core != null) {
             core.removeListener(mListener);
+        }
+
+        super.onPause();
+    }
+
+    private void enableVideoPreviewIfTablet(boolean enable) {
+        if (isTablet()
+                && getResources().getBoolean(R.bool.show_camera_preview_on_dialer_on_tablets)) {
+            Core core = LinphoneManager.getCore();
+            if (enable) {
+                TextureView preview = findViewById(R.id.video_preview);
+                if (preview != null && core != null) {
+                    preview.setVisibility(View.VISIBLE);
+                    core.setNativePreviewWindowId(preview);
+                    core.enableVideoPreview(true);
+                    ImageView changeCamera = findViewById(R.id.video_preview_change_camera);
+                    if (changeCamera != null && core.getVideoDevicesList().length > 1) {
+                        changeCamera.setVisibility(View.VISIBLE);
+                        changeCamera.setOnClickListener(
+                                new View.OnClickListener() {
+                                    @Override
+                                    public void onClick(View v) {
+                                        LinphoneManager.getCallManager().switchCamera();
+                                    }
+                                });
+                    }
+                }
+            } else {
+                core.setNativePreviewWindowId(null);
+                core.enableVideoPreview(false);
+            }
         }
     }